Changes concering adding time registrations.
Trying out new stuff regarding persistence. The command is creating an
instance of a class with the context and what shall be persisted. It is
not how I would usually do it.

Normally that would be a dependency that takes the context into the
constructor and what to save as an argument in the method.

I probably do not like this experimental way for a couple of reasons:

 - The dependency is not shown clearly.
 - Delayed resolving of the dependency makes it less safe. Some IoC
   containers can validate dependencies before constructing the
   resolve root. This would be missed.
 - What is the point of constructing an object which its entire purpose
   is to call one method and then die. It's really just a method call.

The one upside I can see is that I don't have to pass any values through
arguments to the methods. Maybe some of the downsides can be mitigated
by using an abstract factory.

I realize now that I will change this in some way but this commit will
remain for documentation purposes and a way to make my mind up on this
issue.